move and/or jobs to run package
